feat(client): show session logs for sub containers
